retro_frontend: Move frontend state to a global singleton
The public `Frontend` struct now has no members anymore, as it just calls into the implementation singleton.

This leads me to believe that it might be a good idea to maybe embrace that. There are two ways I can imagine right now:

-  not have a Frontend struct at all, but rather a set of free functions that simply make the once_cell (and yes I know once_cell is partially standardized, but the part that I actually want to use, `Lazy`, is *not*.) global API "safer" and hide away the nastiness that libretro makes us do :(
Or:
-  remove the ability to ::new() a Frontend and remove self bindings on the currently exposed functions.

Either way, this would remove the requirement to sanity check quite as much, because the free/static functions don't exactly let you create multiple of a wrapper object, or communicate an assumption which simply isn't true.

The former is probably saner, so I'll probably do that.
