+++ crates/retro_frontend/Cargo.toml
@@ -6,4 +6,5 @@ edition = "2021"
 [dependencies]
 libloading = "0.8.3"
 libretro-sys = "0.1.1"
+once_cell = "1.19.0"
 thiserror = "1.0.61"
