+++ crates/retro_frontend/src/lib.rs
@@ -1,5 +1,7 @@
 //! A libretro frontend as a reusable library crate
 
+mod frontend_impl;
+
 pub mod frontend;
 pub use frontend::*;
